🔨 refactor: 중복 repository 통합
