+++ src/main/java/com/x1/groo/forest/common/domain/repository/DiaryRepository.java
@@ -1,0 +1,4 @@
+package com.x1.groo.forest.common.domain.repository;
+
+public interface DiaryRepository {
+}
